requirements.txt
python-dotenv
langchain
langchain_openai
openai
tiktoken
streamlit
python-nmap
pydantic
pyyaml
pyyaml-env-tag
langchain_chroma